search state
added various mehtods to the search satte such as eveualting the current prioorty of the resources we cna see along with a priority list that will contain all of the viissble resource positions that we saw, these positions are pushed in order taking into account their priority. this will help when transitioning from other states to search based on priroity as it will be able to pick up on what is curreenlty high priorty. so say we are in the retreat state and we see a health pick up we could immidealey trantiosn into the search state which will then go for the pick up due to it being of major or lower priority. Also added methods to get base position while their alive  which can be used for retreating.