added logs for states, adjusted retreat
retreat will need to be tested more but i added capability for the tank to stop for a certain amount of time and check behind it(it stops following the path) allowing us to check behind if needed. also adjusted a few conditions in retreat.  It will also constanlty look behjind itslef when retreating and then when the enemy tank bvecomes null will stop, resources always gathered done in order based on priority. chase and attack now properly transition based on distance along with search state now transitioning based on same distances as attack and chase. added more utility methods to smart tank.